Spostate oltre 1000 righe di codice, 1000 rimanenti.
