+++ BusinessLayer/BL_QuestionManagement.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SchoolGrades
+{
+    internal partial class BusinessLayer
+    {
+    }
+}
