+++ BusinessLayer/BusinessLayer.projitems
@@ -15,7 +15,9 @@
     <Compile Include="$(MSBuildThisFileDirectory)BL_ConstructorsAndGeneralMethods.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_GradesManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_ImageManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)BL_LessonManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_LinkManagement.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)BL_QuestionManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_StudentManagement.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)BL_UserManagement.cs" />
   </ItemGroup>
