+++ DataLayer/DL_QuestionManagement.cs
@@ -1,0 +1,382 @@
+﻿using SchoolGrades.DbClasses;
+using System;
+using System.Collections.Generic;
+using System.Data.Common;
+using System.Text;
+
+namespace SchoolGrades
+{
+    internal partial class DataLayer
+    {
+        private string MakeStringForFilteredQuestionsQuery(List<Tag> Tags, string IdSchoolSubject,
+            string IdQuestionType, Topic QuestionsTopic, bool QueryManyTopics, bool TagsAnd)
+        {
+            string query = "SELECT DISTINCT Questions.idQuestion" +
+                            " FROM Questions";
+            if (Tags != null && Tags.Count > 0)
+            {
+                // must join info from table Questions_Tags
+                string queryTags = " JOIN Questions_Tags ON Questions.idQuestion=Questions_Tags.idQuestion";
+
+                // make an IN clause, useful for both queries
+                string InClause = " WHERE Questions_Tags.idTag IN(";
+                foreach (Tag tag in Tags)
+                {
+                    InClause += "" + tag.IdTag.ToString() + ",";
+                }
+                InClause = InClause.Substring(0, InClause.Length - 1);
+                InClause += ")";
+
+                // (se http://howto.philippkeller.com/2005/04/24/Tags-Database-schemas/, "Toxi" solution)
+                if (!TagsAnd)
+                {
+                    // The tags are evaluated in Union (OR) 
+                    // limits the query only to those questions that have been associated to at least one of the tags in the list
+                    queryTags += InClause;
+                }
+                else
+                {
+                    // The tags are in intersection (AND) 
+                    queryTags += InClause;
+                    queryTags += " GROUP BY Questions.idQuestion";
+                    queryTags += " HAVING COUNT(Questions.idQuestion)=" + Tags.Count;
+                }
+                query += queryTags;
+            }
+            if (IdSchoolSubject != "")
+            {
+                // if we have already added the SQL for tags, we don't need a where
+                if (Tags != null && Tags.Count > 0)
+                    query += " AND idSchoolSubject ='" + SqlVal.SqlString(IdSchoolSubject) + "'";
+                else
+                    query += " WHERE idSchoolSubject ='" + SqlVal.SqlString(IdSchoolSubject) + "'";
+            }
+            if (IdQuestionType != null && IdQuestionType != "")
+            {
+                if (IdSchoolSubject != "" || Tags.Count > 0)
+                    query += " AND idQuestionType ='" + SqlVal.SqlString(IdQuestionType) + "'";
+                else
+                    query += " WHERE idQuestionType ='" + SqlVal.SqlString(IdQuestionType) + "'";
+            }
+            if (QuestionsTopic != null)
+            {
+                if (!QueryManyTopics)
+                {
+                    if (IdSchoolSubject != "" || Tags.Count > 0 || IdQuestionType != "")
+                        query += " AND idTopic=" + QuestionsTopic.Id + "";
+                    else
+                        query += " WHERE idTopic=" + QuestionsTopic.Id + "";
+                }
+                else
+                {
+                    string queryApplicableTopics = "SELECT idTopic FROM Topics" +
+                        " WHERE Topics.leftNode BETWEEN " + QuestionsTopic.LeftNodeOld +
+                        " AND " + QuestionsTopic.RightNodeOld;
+                    // query the passed Topic, plus all its descendants in the tree
+                    if (IdSchoolSubject != "" || Tags.Count > 0 || IdQuestionType != "")
+                        query += " AND Questions.IdTopic IN (" + queryApplicableTopics + ")";
+                    else
+                        query += " WHERE Questions.IdTopic IN (" + queryApplicableTopics + ")";
+                }
+            }
+            return query;
+        }
+
+        internal List<Question> GetAllQuestionsOfATest(int? IdTest)
+        {
+            List<Question> lq = new List<Question>();
+            using (DbConnection conn = Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT *" +
+                    " FROM Questions" +
+                    " JOIN Tests_Questions ON Tests_Questions.IdQuestion=Questions.IdQuestion" +
+                    " WHERE Tests_Questions.idTest=" + IdTest +
+                    ";";
+                DbDataReader dRead;
+                dRead = cmd.ExecuteReader();
+
+                while (dRead.Read())
+                {
+                    Question q = GetQuestionFromRow(dRead);
+                    lq.Add(q);
+                }
+            }
+            return lq;
+        }
+
+        private Question GetQuestionFromRow(DbDataReader Row)
+        {
+            Question q = new Question();
+            q.Difficulty = SafeDb.SafeInt(Row["Difficulty"]);
+            q.IdImage = SafeDb.SafeInt(Row["IdImage"]);
+            q.Duration = SafeDb.SafeInt(Row["Duration"]);
+            q.IdQuestion = SafeDb.SafeInt(Row["IdQuestion"]);
+            q.IdQuestionType = SafeDb.SafeString(Row["IdQuestionType"]);
+            q.IdSchoolSubject = SafeDb.SafeString(Row["IdSchoolSubject"]);
+            //q.IdSubject = SafeDb.SafeInt(Row["IdSubject"]);
+            q.IdTopic = SafeDb.SafeInt(Row["IdTopic"]);
+            q.Image = SafeDb.SafeString(Row["Image"]);
+            q.Text = SafeDb.SafeString(Row["Text"]);
+            q.Weight = SafeDb.SafeDouble(Row["Weight"]);
+            q.NRows = SafeDb.SafeInt(Row["nRows"]);
+            q.IsParamount = SafeDb.SafeInt(Row["isParamount"]);
+            ////////q.IsFixed = SafeDb.SafeBool(Row["isFixed"]);
+
+            return q;
+        }
+
+        internal Question GetQuestionById(int? IdQuestion)
+        {
+            Question question = new Question();
+            if (IdQuestion == 0)
+            {
+                question.IdQuestion = 0;
+                question.Text = "";
+                return question;
+            }
+            using (DbConnection conn = Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT * " +
+                    "FROM Questions " +
+                    "WHERE idQuestion=" + IdQuestion.ToString() +
+                    ";";
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read()) // we cycle even if we need just a row, to check if it exists
+                {
+                    question.IdQuestion = (int)dRead["idQuestion"];
+                    question.IdQuestionType = dRead["idQuestionType"].ToString();
+                    question.Text = dRead["text"].ToString();
+                    question.QuestionImage = dRead["image"].ToString();
+                    question.Difficulty = (int)dRead["difficulty"];
+                    question.Duration = (int)dRead["duration"];
+                    question.IdSchoolSubject = dRead["idSchoolSubject"].ToString();
+                    question.IdTopic = (int)dRead["idTopic"];
+                    question.Image = dRead["image"].ToString();
+                    question.Weight = (double)dRead["weight"];
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return question;
+        }
+
+        internal void SaveQuestion(Question Question)
+        {
+            using (DbConnection conn = Connect())
+            {
+                string imageNoHome = Question.Image;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "UPDATE Questions " +
+                    "SET idQuestionType='" + SqlVal.SqlString(Question.IdQuestionType) + "' " +
+                     ", idSchoolSubject='" + SqlVal.SqlString(Question.IdSchoolSubject) + "' " +
+                     //", idSubject=" + Question.IdSubject + " " +
+                     ", idSchoolSubject='" + Question.IdSchoolSubject + "'" +
+                     ", idTopic=" + Question.IdTopic + " " +
+                     ", duration=" + Question.Duration + " " +
+                     ", difficulty=" + Question.Difficulty + " " +
+                     ", text='" + SqlVal.SqlString(Question.Text) + "' " +
+                     ", image='" + SqlVal.SqlString(imageNoHome) + "' " +
+                     ", weight=" + SqlVal.SqlDouble(Question.Weight.ToString()) + " " +
+                    "WHERE idQuestion=" + Question.IdQuestion +
+                    ";";
+                cmd.ExecuteNonQuery();
+                // !!!!TODO sistemare le risposte
+                // !!!!TODO gestire i tag
+                cmd.Dispose();
+            }
+        }
+
+        internal void AddQuestionToTest(Test Test, Question Question)
+        {
+            using (DbConnection conn = Connect())
+            {
+                // get the code of the previous photo
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Tests_Questions" +
+                    " (IdTest, IdQuestion)" +
+                    " Values" +
+                    " (" + Test.IdTest + "," + Question.IdQuestion + ")" +
+                    "; ";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+        internal List<QuestionType> GetListQuestionTypes(bool IncludeANullObject)
+        {
+            List<QuestionType> l = new List<QuestionType>();
+            if (IncludeANullObject)
+            {
+                QuestionType qt = new QuestionType();
+                qt.IdQuestionType = "";
+                l.Add(qt);
+            }
+            using (DbConnection conn = Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT * FROM QuestionTypes;";
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    QuestionType o = new QuestionType();
+                    o.Name = (string)dRead["Name"];
+                    o.IdQuestionType = (string)dRead["IdQuestionType"];
+                    o.Desc = (string)dRead["Desc"];
+                    o.IdQuestionType = (string)dRead["IdQuestionType"];
+                    l.Add(o);
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+                return l;
+            }
+        }
+
+        internal Question CreateNewVoidQuestion()
+        {
+            // trova una chiave da assegnare alla nuova domanda
+            Question q = new Question();
+            q.IdQuestion = NextKey("Questions", "idQuestion");
+            using (DbConnection conn = Connect())
+            {
+                string imageSenzaHome = q.Image;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Questions " +
+                    "(idQuestion) " +
+                    "Values (" + q.IdQuestion + ")" +
+                     ";";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+            SaveQuestion(q);
+            return q;
+        }
+
+        /// <summary>
+        /// gets the questions regarding the topics taught to the class that 
+        /// haven't been made to the student yet. 
+        /// Includes also the questions tha do not have a topic 
+        /// </summary>
+        /// <param name="Class"></param>
+        /// <param name="Student"></param>
+        /// <param name="Subject"></param>
+        /// <returns></returns>
+        internal List<Question> GetFilteredQuestionsNotAsked(Student Student, Class Class,
+            SchoolSubject Subject, string IdQuestionType, List<Tag> Tags, Topic Topic,
+            bool QueryManyTopics, bool TagsAnd, string SearchString,
+            DateTime DateFrom, DateTime DateTo)
+        {
+            List<Question> lq = new List<Question>();
+            string filteredQuestions;
+
+            // first part of the query: selection of the interesting fields in Questions
+            string query = "SELECT Questions.IdQuestion,Questions.text,Questions.idSchoolSubject,Questions.idQuestionType" +
+                ",Questions.weight,Questions.duration,Questions.difficulty,Questions.image,Questions.idTopic" +
+                " FROM Questions";
+            // add the WHERE clauses
+            // if the search string is present, then it must be in the searched field 
+            if (SearchString != "")
+            {
+                query += " WHERE Questions.text LIKE('%" + SqlVal.SqlString(SearchString) + "%')" +
+                    "AND (";
+            }
+            if (Subject != null)
+                filteredQuestions = MakeStringForFilteredQuestionsQuery(Tags, Subject.IdSchoolSubject, IdQuestionType,
+                    Topic, QueryManyTopics, TagsAnd);
+            else
+                filteredQuestions = MakeStringForFilteredQuestionsQuery(Tags, "", IdQuestionType,
+                    Topic, QueryManyTopics, TagsAnd);
+
+            string questionsAlreadyMade = "";
+            if (Student != null)
+            {
+                questionsAlreadyMade = "SELECT Questions.idQuestion" +
+                    " FROM Questions" +
+                    " JOIN Grades ON Questions.idQuestion=Grades.idQuestion" +
+                    " JOIN Students ON Students.idStudent=Grades.IdStudent" +
+                    " WHERE Students.idStudent=" + Student.IdStudent +
+                    " AND Grades.idSchoolYear='" + Class.SchoolYear + "'";
+            }
+            string questionsTopicsMade = "";
+            if (Class != null && Subject != null)
+            {
+                // questions made to the class in every time ever 
+                questionsTopicsMade = "SELECT Questions.idQuestion" +
+                    " FROM Questions" +
+                    " JOIN Lessons_Topics ON Questions.idTopic=Lessons_Topics.idTopic" +
+                    " JOIN Lessons ON Lessons_Topics.idLesson=Lessons.idLesson" +
+                    " JOIN Classes ON Classes.idClass=Lessons.idClass" +
+                    " WHERE Classes.idClass=" + Class.IdClass +
+                    " AND (Questions.idSchoolSubject='" + Subject.IdSchoolSubject + "'" +
+                    " OR Questions.idSchoolSubject='' OR Questions.idSchoolSubject=NULL)";
+                if (DateFrom != Commons.DateNull)
+                    questionsTopicsMade += " AND (Lessons.Date BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) + ")";
+                // PART of the final query that extracts the Ids of the questions already made 
+                questionsTopicsMade = " Questions.idQuestion IN(" + questionsTopicsMade + ")";
+            }
+
+            if (questionsAlreadyMade != "")
+            {
+                // take only questions already made 
+                if (SearchString == "")
+                {
+                    query += " WHERE Questions.idQuestion NOT IN(" + questionsAlreadyMade + ")";
+                }
+                else
+                {
+                    query += " Questions.idQuestion NOT IN(" + questionsAlreadyMade + ")";
+                }
+            }
+            if (filteredQuestions != "")
+            {
+                if (questionsAlreadyMade != "" || SearchString != "")
+                {
+                    query += " AND Questions.idQuestion IN(" + filteredQuestions + ")";
+                }
+                else
+                {
+                    query += " WHERE Questions.idQuestion IN(" + filteredQuestions + ")";
+                }
+            }
+            query += " OR Questions.idTopic IS NULL OR Questions.idTopic = ''";
+            if (SearchString != "")
+                query += ")";
+
+            query += " ORDER BY Questions.weight;";
+
+            using (DbConnection conn = Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+
+                cmd.CommandText = query;
+
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read()) // 
+                {
+                    Question questionForList = new Question();
+
+                    questionForList.Difficulty = (int)dRead["difficulty"];
+                    questionForList.Duration = (int)dRead["duration"];
+                    questionForList.IdQuestion = (int)dRead["idQuestion"];
+                    questionForList.IdQuestionType = dRead["idQuestionType"].ToString();
+                    questionForList.IdSchoolSubject = dRead["idSchoolSubject"].ToString();
+                    //q.idSubject = (int)dRead["idSubject"];
+                    questionForList.IdTopic = (int)dRead["idTopic"];
+                    questionForList.Image = dRead["image"].ToString();
+                    questionForList.QuestionImage = dRead["image"].ToString();
+                    questionForList.Text = dRead["text"].ToString();
+                    questionForList.Weight = (double)dRead["weight"];
+
+                    lq.Add(questionForList);
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return lq;
+        }
+    }
+}
