+++ DbClasses/DbAndBusiness.cs
@@ -66,23 +66,6 @@ namespace SchoolGrades.DbClasses
             }
         }
 
-        //internal void UpdatePathStartLinkOfClass(Class currentClass, string text)
-        //{
-        //    // !!!! currently not used, because pathStartLink field does not exist yet in the database !!!!
-        //    using (DbConnection conn = dl.Connect())
-        //    {
-        //        DbCommand cmd = conn.CreateCommand();
-
-        //        cmd.CommandText = "UPDATE Classes" +
-        //                   " Set" +
-        //                   " pathStartLink='" + text + "'" +
-        //                   " WHERE IdClass=" + currentClass.IdClass +
-        //                   ";";
-        //        cmd.ExecuteNonQuery();
-        //        cmd.Dispose();
-        //    }
-        //}
-
         internal void SaveTableOnCvs(DataTable Table, string FileName)
         {
             string fileContent = "";
@@ -194,34 +177,6 @@ namespace SchoolGrades.DbClasses
                 return NamePath;
             }
         }
-
-        //internal void AddLinkToOldPhoto(int? IdStudent, string IdPreviousSchoolYear, string IdNextSchoolYear)
-        //{
-        //    using (DbConnection conn = dl.Connect())
-        //    {
-        //        // get the code of the previous photo
-        //        DbCommand cmd = conn.CreateCommand();
-        //        cmd.CommandText = "SELECT idStudentsPhoto" +
-        //            " FROM StudentsPhotos_Students" +
-        //            " WHERE idSchoolYear='" + IdPreviousSchoolYear + "'" +
-        //            " AND StudentsPhotos_Students.idStudent = " + IdStudent + "; ";
-        //        int? idStudentsPhoto = (int?)cmd.ExecuteScalar();
-        //        if (idStudentsPhoto != null)
-        //        {
-        //            // add link to old photo
-        //            cmd = conn.CreateCommand();
-        //            cmd.CommandText = "INSERT INTO StudentsPhotos_Students " +
-        //            "(idStudent, idStudentsPhoto, idSchoolYear) " +
-        //            "Values (" +
-        //            "" + IdStudent + "" +
-        //            "," + idStudentsPhoto + "" +
-        //            ",'" + IdNextSchoolYear + "'" +
-        //            ");";
-        //            cmd.ExecuteNonQuery();
-        //        }
-        //        cmd.Dispose();
-        //    }
-        //}
 
         internal bool IsTopicAlreadyTaught(Topic Topic)
         {
@@ -362,51 +317,6 @@ namespace SchoolGrades.DbClasses
             return d;
         }
 
-        internal DataTable GetSubGradesOfGrade(int? IdGrade)
-        {
-            DataTable t;
-            using (DbConnection conn = dl.Connect())
-            {
-                string query = "SELECT datetime(Grades.timestamp),Questions.text,Grades.value," +
-                    " Grades.weight,Grades.cncFactor,Grades.idGradeParent" +
-                    " FROM Grades" +
-                    " JOIN Questions ON Grades.idQuestion=Questions.idQuestion" +
-                    " WHERE Grades.idGradeParent =" + IdGrade +
-                    " ORDER BY Grades.timestamp;";
-
-                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                DataSet DSet = new DataSet("OpenMicroGrades");
-                DAdapt.Fill(DSet);
-                t = DSet.Tables[0];
-
-                DAdapt.Dispose();
-                DSet.Dispose();
-            }
-            return t;
-        }
-
-        
-
-        internal Question CreateNewVoidQuestion()
-        {
-            // trova una chiave da assegnare alla nuova domanda
-            Question q = new Question();
-            q.IdQuestion = dl.NextKey("Questions", "idQuestion");
-            using (DbConnection conn = dl.Connect())
-            {
-                string imageSenzaHome = q.Image;
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "INSERT INTO Questions " +
-                    "(idQuestion) " +
-                    "Values (" + q.IdQuestion + ")" +
-                     ";";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-            SaveQuestion(q);
-            return q;
-        }
-
         internal SchoolSubject GetSchoolSubject(string IdSchoolSubject)
         {
             SchoolSubject subject = new SchoolSubject();
@@ -469,292 +379,7 @@ namespace SchoolGrades.DbClasses
             cmd.Dispose();
             dRead.Dispose();
             return lss;
-        }
-
-        internal Question GetQuestionById(int? IdQuestion)
-        {
-            Question question = new Question();
-            if (IdQuestion == 0)
-            {
-                question.IdQuestion = 0;
-                question.Text = ""; 
-                return question;
-            }
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT * " +
-                    "FROM Questions " +
-                    "WHERE idQuestion=" + IdQuestion.ToString() +
-                    ";";
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read()) // we cycle even if we need just a row, to check if it exists
-                {
-                    question.IdQuestion = (int)dRead["idQuestion"];
-                    question.IdQuestionType = dRead["idQuestionType"].ToString();
-                    question.Text = dRead["text"].ToString();
-                    question.QuestionImage = dRead["image"].ToString();
-                    question.Difficulty = (int)dRead["difficulty"];
-                    question.Duration = (int)dRead["duration"];
-                    question.IdSchoolSubject = dRead["idSchoolSubject"].ToString();
-                    question.IdTopic = (int)dRead["idTopic"];
-                    question.Image = dRead["image"].ToString();
-                    question.Weight = (double)dRead["weight"];
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return question;
-        }
-
-        internal void SaveQuestion(Question Question)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                string imageNoHome = Question.Image;
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "UPDATE Questions " +
-                    "SET idQuestionType='" + SqlVal.SqlString(Question.IdQuestionType) + "' " +
-                     ", idSchoolSubject='" + SqlVal.SqlString(Question.IdSchoolSubject) + "' " +
-                     //", idSubject=" + Question.IdSubject + " " +
-                     ", idSchoolSubject='" + Question.IdSchoolSubject + "'" +
-                     ", idTopic=" + Question.IdTopic + " " +
-                     ", duration=" + Question.Duration + " " +
-                     ", difficulty=" + Question.Difficulty + " " +
-                     ", text='" + SqlVal.SqlString(Question.Text) + "' " +
-                     ", image='" + SqlVal.SqlString(imageNoHome) + "' " +
-                     ", weight=" + SqlVal.SqlDouble(Question.Weight.ToString()) + " " +
-                    "WHERE idQuestion=" + Question.IdQuestion +
-                    ";";
-                cmd.ExecuteNonQuery();
-                // !!!!TODO sistemare le risposte
-                // !!!!TODO gestire i tag
-                cmd.Dispose();
-            }
-        }
-
-        internal DataTable GetFilteredQuestions(List<Tag> Tags, string IdSchoolSubject,
-            string IdQuestionType, Topic QuestionsTopic, bool QueryManyTopics, bool TagsAnd)
-        {
-            DataTable t;
-
-            string subquery = MakeStringForFilteredQuestionsQuery(Tags, IdSchoolSubject, IdQuestionType, QuestionsTopic, QueryManyTopics, TagsAnd);
-            string query = "SELECT Questions.IdQuestion,Questions.text,Questions.idSchoolSubject,Questions.idQuestionType" +
-                            ",Questions.weight,Questions.duration,Questions.difficulty,Questions.image,Questions.idTopic" +
-                            " FROM Questions" +
-                            " WHERE Questions.IdQuestion IN(" + subquery + ")" +
-                            " ORDER BY Questions.IdQuestion;";
-            DataSet DSet;
-            using (DbConnection conn = dl.Connect())
-            {
-                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                DSet = new DataSet("FilteredQuestions");
-                DAdapt.Fill(DSet);
-                DAdapt.Dispose();
-                t = DSet.Tables[0];
-                DSet.Dispose();
-            }
-            return t;
-        }
-
-        /// <summary>
-        /// gets the questions regarding the topics taught to the class that 
-        /// haven't been made to the student yet. 
-        /// Includes also the questions tha do not have a topic 
-        /// </summary>
-        /// <param name="Class"></param>
-        /// <param name="Student"></param>
-        /// <param name="Subject"></param>
-        /// <returns></returns>
-        internal List<Question> GetFilteredQuestionsNotAsked(Student Student, Class Class,
-            SchoolSubject Subject, string IdQuestionType, List<Tag> Tags, Topic Topic,
-            bool QueryManyTopics, bool TagsAnd, string SearchString,
-            DateTime DateFrom, DateTime DateTo)
-        {
-            List<Question> lq = new List<Question>();
-            string filteredQuestions;
-
-            // first part of the query: selection of the interesting fields in Questions
-            string query = "SELECT Questions.IdQuestion,Questions.text,Questions.idSchoolSubject,Questions.idQuestionType" +
-                ",Questions.weight,Questions.duration,Questions.difficulty,Questions.image,Questions.idTopic" +
-                " FROM Questions";
-            // add the WHERE clauses
-            // if the search string is present, then it must be in the searched field 
-            if (SearchString != "")
-            {
-                query += " WHERE Questions.text LIKE('%" + SqlVal.SqlString(SearchString) + "%')" +
-                    "AND (";
-            }
-            if (Subject != null)
-                filteredQuestions = MakeStringForFilteredQuestionsQuery(Tags, Subject.IdSchoolSubject, IdQuestionType,
-                    Topic, QueryManyTopics, TagsAnd);
-            else
-                filteredQuestions = MakeStringForFilteredQuestionsQuery(Tags, "", IdQuestionType,
-                    Topic, QueryManyTopics, TagsAnd);
-
-            string questionsAlreadyMade = "";
-            if (Student != null)
-            {
-                questionsAlreadyMade = "SELECT Questions.idQuestion" +
-                    " FROM Questions" +
-                    " JOIN Grades ON Questions.idQuestion=Grades.idQuestion" +
-                    " JOIN Students ON Students.idStudent=Grades.IdStudent" +
-                    " WHERE Students.idStudent=" + Student.IdStudent +
-                    " AND Grades.idSchoolYear='" + Class.SchoolYear + "'";
-            }
-            string questionsTopicsMade = "";
-            if (Class != null && Subject != null)
-            {
-                // questions made to the class in every time ever 
-                questionsTopicsMade = "SELECT Questions.idQuestion" +
-                    " FROM Questions" +
-                    " JOIN Lessons_Topics ON Questions.idTopic=Lessons_Topics.idTopic" +
-                    " JOIN Lessons ON Lessons_Topics.idLesson=Lessons.idLesson" +
-                    " JOIN Classes ON Classes.idClass=Lessons.idClass" +
-                    " WHERE Classes.idClass=" + Class.IdClass +
-                    " AND (Questions.idSchoolSubject='" + Subject.IdSchoolSubject + "'" +
-                    " OR Questions.idSchoolSubject='' OR Questions.idSchoolSubject=NULL)";
-                if (DateFrom != Commons.DateNull)
-                    questionsTopicsMade += " AND (Lessons.Date BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) + ")";
-                // PART of the final query that extracts the Ids of the questions already made 
-                questionsTopicsMade = " Questions.idQuestion IN(" + questionsTopicsMade + ")";
-            }
-
-            if (questionsAlreadyMade != "")
-            {
-                // take only questions already made 
-                if (SearchString == "")
-                {
-                    query += " WHERE Questions.idQuestion NOT IN(" + questionsAlreadyMade + ")";
-                }
-                else
-                {
-                    query += " Questions.idQuestion NOT IN(" + questionsAlreadyMade + ")";
-                }
-            }
-            if (filteredQuestions != "")
-            {
-                if (questionsAlreadyMade != "" || SearchString != "")
-                {
-                    query += " AND Questions.idQuestion IN(" + filteredQuestions + ")";
-                }
-                else
-                {
-                    query += " WHERE Questions.idQuestion IN(" + filteredQuestions + ")";
-                }
-            }
-            query += " OR Questions.idTopic IS NULL OR Questions.idTopic = ''";
-            if (SearchString != "")
-                query += ")"; 
-
-            query += " ORDER BY Questions.weight;";
-
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-
-                cmd.CommandText = query;
-
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read()) // 
-                {
-                    Question questionForList = new Question();
-
-                    questionForList.Difficulty = (int)dRead["difficulty"];
-                    questionForList.Duration = (int)dRead["duration"];
-                    questionForList.IdQuestion = (int)dRead["idQuestion"];
-                    questionForList.IdQuestionType = dRead["idQuestionType"].ToString();
-                    questionForList.IdSchoolSubject = dRead["idSchoolSubject"].ToString();
-                    //q.idSubject = (int)dRead["idSubject"];
-                    questionForList.IdTopic = (int)dRead["idTopic"];
-                    questionForList.Image = dRead["image"].ToString();
-                    questionForList.QuestionImage = dRead["image"].ToString();
-                    questionForList.Text = dRead["text"].ToString();
-                    questionForList.Weight = (double)dRead["weight"];
-
-                    lq.Add(questionForList);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return lq;
-        }
-
-        private string MakeStringForFilteredQuestionsQuery(List<Tag> Tags, string IdSchoolSubject,
-            string IdQuestionType, Topic QuestionsTopic, bool QueryManyTopics, bool TagsAnd)
-        {
-            string query = "SELECT DISTINCT Questions.idQuestion" +
-                            " FROM Questions";
-            if (Tags != null && Tags.Count > 0)
-            {
-                // must join info from table Questions_Tags
-                string queryTags = " JOIN Questions_Tags ON Questions.idQuestion=Questions_Tags.idQuestion";
-
-                // make an IN clause, useful for both queries
-                string InClause = " WHERE Questions_Tags.idTag IN(";
-                foreach (Tag tag in Tags)
-                {
-                    InClause += "" + tag.IdTag.ToString() + ",";
-                }
-                InClause = InClause.Substring(0, InClause.Length - 1);
-                InClause += ")";
-
-                // (se http://howto.philippkeller.com/2005/04/24/Tags-Database-schemas/, "Toxi" solution)
-                if (!TagsAnd)
-                {
-                    // The tags are evaluated in Union (OR) 
-                    // limits the query only to those questions that have been associated to at least one of the tags in the list
-                    queryTags += InClause;
-                }
-                else
-                {
-                    // The tags are in intersection (AND) 
-                    queryTags += InClause;
-                    queryTags += " GROUP BY Questions.idQuestion";
-                    queryTags += " HAVING COUNT(Questions.idQuestion)=" + Tags.Count;
-                }
-                query += queryTags;
-            }
-            if (IdSchoolSubject != "")
-            {
-                // if we have already added the SQL for tags, we don't need a where
-                if (Tags != null && Tags.Count > 0)
-                    query += " AND idSchoolSubject ='" + SqlVal.SqlString(IdSchoolSubject) + "'";
-                else
-                    query += " WHERE idSchoolSubject ='" + SqlVal.SqlString(IdSchoolSubject) + "'";
-            }
-            if (IdQuestionType != null && IdQuestionType != "")
-            {
-                if (IdSchoolSubject != "" || Tags.Count > 0)
-                    query += " AND idQuestionType ='" + SqlVal.SqlString(IdQuestionType) + "'";
-                else
-                    query += " WHERE idQuestionType ='" + SqlVal.SqlString(IdQuestionType) + "'";
-            }
-            if (QuestionsTopic != null)
-            {
-                if (!QueryManyTopics)
-                {
-                    if (IdSchoolSubject != "" || Tags.Count > 0 || IdQuestionType != "")
-                        query += " AND idTopic=" + QuestionsTopic.Id + "";
-                    else
-                        query += " WHERE idTopic=" + QuestionsTopic.Id + "";
-                }
-                else
-                {
-                    string queryApplicableTopics = "SELECT idTopic FROM Topics" +
-                        " WHERE Topics.leftNode BETWEEN " + QuestionsTopic.LeftNodeOld +
-                        " AND " + QuestionsTopic.RightNodeOld;
-                    // query the passed Topic, plus all its descendants in the tree
-                    if (IdSchoolSubject != "" || Tags.Count > 0 || IdQuestionType != "")
-                        query += " AND Questions.IdTopic IN (" + queryApplicableTopics + ")";
-                    else
-                        query += " WHERE Questions.IdTopic IN (" + queryApplicableTopics + ")";
-                }
-            }
-            return query;
-        }
+        }      
 
         internal void GetLookupTable(string Table, ref DataSet DSet, ref DataAdapter DAdapt)
         {
@@ -908,373 +533,6 @@ namespace SchoolGrades.DbClasses
                     IdTag +
                     ");";
                 cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-        
-
-        
-       
-
-        internal List<Answer> GetAnswersOfAQuestion(int? idQuestion)
-        {
-            List<Answer> l = new List<Answer>();
-            DbDataReader dRead;
-            DbCommand cmd;
-            using (DbConnection conn = dl.Connect())
-            {
-                string query = "SELECT *" +
-                    " FROM Answers" +
-                    " WHERE idQuestion=" + idQuestion +
-                    " ORDER BY showingOrder;";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Answer t = new Answer();
-                    t.IdAnswer = (int)dRead["idAnswer"];
-                    t.IdQuestion = (int)dRead["idQuestion"];
-                    t.ShowingOrder = (int)dRead["showingOrder"];
-                    t.Text = (string)dRead["text"];
-                    t.IdAnswer = (int)dRead["idAnswer"];
-                    t.ErrorCost = (int)dRead["errorCost"];
-                    t.IsCorrect = SafeDb.SafeBool(dRead["isCorrect"]);
-                    t.IsOpenAnswer = SafeDb.SafeBool(dRead["isOpenAnswer"]);
-
-                    l.Add(t);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return l;
-        }
-        
-        internal List<string> GetCaptionsOfThisImage(string FileName)
-        {
-            List<string> captions = new List<string>();
-
-            DbDataReader dRead;
-            DbCommand cmd;
-            using (DbConnection conn = dl.Connect())
-            {
-                cmd = conn.CreateCommand();
-                string query;
-                query = "SELECT Caption FROM Images" +
-                        " WHERE imagePath LIKE '%" + FileName + "%'";
-                query += ";";
-                cmd.CommandText = query;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    captions.Add((string)dRead["Caption"]);
-                }
-                cmd.Dispose();
-                dRead.Dispose();
-            }
-            return captions;
-        }
-
-        internal void BackupAllStudentsDataTsv()
-        {
-            BackupTableTsv("Students");
-            BackupTableTsv("StudentsPhotos");
-            BackupTableTsv("StudentsPhotos_Students");
-            BackupTableTsv("Classes_Students");
-            BackupTableTsv("Grades");
-        }
-
-        internal void BackupAllStudentsDataXml()
-        {
-            BackupTableXml("Students");
-            BackupTableXml("StudentsPhotos");
-            BackupTableXml("StudentsPhotos_Students");
-            BackupTableXml("Classes_Students");
-            BackupTableXml("Grades");
-        }
-
-        internal void RestoreAllStudentsDataTsv(bool MustErase)
-        {
-            RestoreTableTsv("Students", MustErase);
-            RestoreTableTsv("StudentsPhotos", MustErase);
-            RestoreTableTsv("StudentsPhotos_Students", MustErase);
-            RestoreTableTsv("Classes_Students", MustErase);
-            RestoreTableTsv("Grades", MustErase);
-        }
-
-        internal void RestoreAllStudentsDataXml(bool MustErase)
-        {
-            RestoreTableXml("Students", MustErase);
-            RestoreTableXml("StudentsPhotos", MustErase);
-            RestoreTableXml("StudentsPhotos_Students", MustErase);
-            RestoreTableXml("Classes_Students", MustErase);
-            RestoreTableXml("Grades", MustErase);
-        }
-
-        internal void BackupTableTsv(string TableName)
-        {
-            DbDataReader dRead;
-            DbCommand cmd;
-            string fileContent = "";
-
-            using (DbConnection conn = dl.Connect())
-            {
-                string query = "SELECT *" +
-                    " FROM " + TableName + " ";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                dRead = cmd.ExecuteReader();
-                int y = 0;
-                while (dRead.Read())
-                {
-                    // field names only in first row 
-                    if (y == 0)
-                    {
-                        string types = "";
-                        for (int i = 0; i < dRead.FieldCount; i++)
-                        {
-                            fileContent += "\"" + dRead.GetName(i) + "\"\t";
-                            types += "\"" + SafeDb.SafeString(dRead.GetDataTypeName(i)) + "\"\t";
-                        }
-                        fileContent = fileContent.Substring(0, fileContent.Length - 1) + "\r\n";
-                        fileContent += types.Substring(0, types.Length - 1) + "\r\n";
-                    }
-                    // field values
-                    string values = "";
-                    if (dRead.GetValue(0) != null)
-                    {
-                        Console.Write(dRead.GetValue(0));
-                        for (int i = 0; i < dRead.FieldCount; i++)
-                        {
-                            values += "\"" + SafeDb.SafeString(dRead.GetValue(i).ToString()) + "\"\t";
-                        }
-                        fileContent += values.Substring(0, values.Length - 1) + "\r\n";
-                    }
-                    else
-                    {
-
-                    }
-                    y++;
-                }
-                TextFile.StringToFile(Commons.PathDatabase + "\\" + TableName + ".tsv", fileContent, false);
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-        }
-
-        internal void BackupTableXml(string TableName)
-        {
-            DataAdapter dAdapt;
-            DataSet dSet = new DataSet();
-            DataTable t;
-            string query = "SELECT *" +
-                    " FROM " + TableName + ";";
-
-            using (DbConnection conn = dl.Connect())
-            {
-                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                dSet = new DataSet("GetTable");
-                dAdapt.Fill(dSet);
-                t = dSet.Tables[0];
-
-                t.WriteXml(Commons.PathDatabase + "\\" + TableName + ".xml",
-                    XmlWriteMode.WriteSchema);
-
-                dAdapt.Dispose();
-                dSet.Dispose();
-            }
-        }
-
-        internal void RestoreTableTsv(string TableName, bool EraseBefore)
-        {
-            List<string> fieldNames;
-            List<string> fieldTypes = new List<string>();
-            //string[,] dati = FileDiTesto.FileInMatrice(Commons.PathDatabase +
-            //    "\\" + TableName + ".tsv", '\t',
-            //    out fieldsNames, out fieldTypes);
-            string dati = TextFile.FileToString(Commons.PathDatabase +
-                "\\" + TableName + ".tsv");
-            if (dati is null)
-                return;
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                if (EraseBefore)
-                {
-                    // first: erases existing rows in the table
-                    cmd.CommandText += "DELETE FROM " + TableName + ";";
-                    cmd.ExecuteNonQuery();
-                }
-                else
-                {
-                    throw new Exception("Append of table records to an existing table id not implemented yet");
-                    //return; 
-                }
-                string fieldsString = " (";
-                string valuesString;
-                int fieldsCount = 0;
-
-                int index = 0;
-                string fieldName = "";
-                while (index < dati.Length)
-                {
-                    // parse first line: field names
-                    fieldNames = new List<string>();
-                    do
-                    {
-                        if (dati[index++] != '\"')
-                            return; // error! 
-                        fieldName = "";
-                        while (dati[index] != '\"')
-                        {
-                            fieldName += dati[index++];
-                        }
-                        fieldNames.Add(fieldName);
-                        fieldsString += fieldName + ",";
-                        fieldsCount++;
-                        if (dati[++index] != '\t' && dati[index] != '\r')
-                            return; // ERROR!
-                    } while (dati[++index] != '\r');
-                    index++; // void line feed
-
-                    // parse second line: field types
-                    string fieldType = "";
-                    while (dati[index] != '\r')
-                    {
-                        while (dati[index] != '\"')
-                        {
-                            fieldType += dati[index++];
-                        }
-                        fieldTypes.Add(fieldType);
-                        fieldsString += fieldName + ",";
-                        fieldsCount++;
-                    }
-                    index++; // void line feed
-
-                    // parse the rest of the rows: values
-                    string fieldValue = "";
-                    while (dati[index] != '\r')
-                    {
-                        while (dati[index] != '\"')
-                        {
-                            fieldType += dati[index++];
-                        }
-                        fieldTypes.Add(fieldType);
-                        fieldsString += fieldName + ",";
-                        fieldsCount++;
-                    }
-                }
-                //for (int col = 0; col < dati.GetLength(1); col++)
-                //{
-                //    if (fieldNames[col] != "")
-                //    {
-                //        fieldsString += fieldNames[col] + ",";
-                //        fieldsCount++; 
-                //    }
-                //}
-                //fieldsString = fieldsString.Substring(0, fieldsString.Length - 1);
-                //fieldsString += ")";
-                //for (int row = 0; row < dati.GetLength(0); row++)
-                //{
-                //    valuesString = " Values (";
-                //    for (int col = 0; col < fieldsCount; col++)
-                //    {
-                //        if (fieldNames[col] != "")
-                //        {
-                //            if (fieldTypes[col].IndexOf("VARCHAR") >= 0)
-                //                valuesString += "'" + SqlVal.SqlString(dati[row, col]) + "',";
-                //            else if (fieldTypes[col].IndexOf("INT") >= 0)
-                //                valuesString +=  SqlVal.SqlInt(dati[row, col]) + ",";
-                //            else if (fieldTypes[col].IndexOf("REAL") >= 0)
-                //                valuesString += SqlFloat(dati[row, col]) + ",";
-                //            else if (fieldTypes[col].IndexOf("FLOAT") >= 0)
-                //                valuesString += SqlFloat(dati[row, col]) + ",";
-                //            else if (fieldTypes[col].IndexOf("DATE") >= 0)
-                //                valuesString += SqlVal.SqlDate(dati[row, col]) + ",";
-                //        }
-                //    }
-                //    valuesString = valuesString.Substring(0, valuesString.Length - 1);
-                //    valuesString += ")";
-                //    cmd.CommandText = "INSERT INTO " + TableName +
-                //                fieldsString +
-                //                valuesString;
-                //    //" WHERE " + fieldsNames[0] + "=";
-                //    //if (fieldTypes[0].IndexOf("VARCHAR") >= 0)
-                //    //    cmd.CommandText += "'" + StringSql(dati[row, 0]) + "'";
-                //    //else
-                //    //    cmd.CommandText += StringSql(dati[row, 0]);
-                //    cmd.CommandText += ";";
-                //    cmd.ExecuteNonQuery();
-                //}
-                //cmd.Dispose();
-            }
-        }
-
-        internal void RestoreTableXml(string TableName, bool EraseBefore)
-        {
-            DataSet dSet = new DataSet();
-            DataTable t = null;
-            dSet.ReadXml(Commons.PathDatabase + "\\" + TableName + ".xml", XmlReadMode.ReadSchema);
-            t = dSet.Tables[0];
-            if (t.Rows.Count == 0)
-                return;
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd;
-                cmd = conn.CreateCommand();
-                if (EraseBefore)
-                {
-                    cmd.CommandText = "DELETE FROM " + TableName + ";";
-                    cmd.ExecuteNonQuery();
-                }
-                cmd.CommandText = "INSERT INTO " + TableName + "(";
-                // column names
-                DataRow r = t.Rows[0];
-                foreach (DataColumn c in t.Columns)
-                {
-                    cmd.CommandText += c.ColumnName + ",";
-                }
-                cmd.CommandText = cmd.CommandText.Substring(0, cmd.CommandText.Length - 1);
-                cmd.CommandText += ")VALUES";
-                // row values
-                foreach (DataRow row in t.Rows)
-                {
-                    cmd.CommandText += "(";
-                    foreach (DataColumn c in t.Columns)
-                    {
-                        switch (Type.GetTypeCode(c.DataType))
-                        {
-                            case TypeCode.String:
-                            case TypeCode.Char:
-                                {
-                                    cmd.CommandText += "'" + SqlVal.SqlString(row[c.ColumnName].ToString()) + "',";
-                                    break;
-                                };
-                            case TypeCode.DateTime:
-                                {
-                                    DateTime? d = SafeDb.SafeDateTime(row[c.ColumnName]);
-                                    cmd.CommandText += "'" +
-                                        ((DateTime)(d)).ToString("yyyy-MM-dd_HH.mm.ss") + "',";
-                                    break;
-                                }
-                            default:
-                                {
-                                    if (!(row[c.ColumnName] is DBNull))
-                                        cmd.CommandText += row[c.ColumnName] + ",";
-                                    else
-                                        cmd.CommandText += "0,";
-                                    break;
-                                }
-                        }
-                    }
-                    cmd.CommandText = cmd.CommandText.Substring(0, cmd.CommandText.Length - 1);
-                    cmd.CommandText += "),";
-                }
-                cmd.CommandText = cmd.CommandText.Substring(0, cmd.CommandText.Length - 1);
-                cmd.CommandText += ";";
-                cmd.ExecuteNonQuery();
-                dSet.Dispose();
-                t.Dispose();
                 cmd.Dispose();
             }
         }
@@ -1586,1212 +844,8 @@ namespace SchoolGrades.DbClasses
             }
             return Class.PathRestrictedApplication;
         }
-        internal string CreateDemoDatabase(string newDatabaseFullName, Class Class1, Class Class2)
-        {
-            DbCommand cmd;
 
-            File.Copy(Commons.PathAndFileDatabase, newDatabaseFullName);
-
-            // local instance of a DataLayer to operate on a second database 
-            DataLayer newDatabaseDl = new DataLayer(newDatabaseFullName);
-
-            // erase all the data of the students of other classes
-            using (DbConnection conn = newDatabaseDl.Connect()) // connect to the new database, just copied
-            {
-                cmd = conn.CreateCommand();
-
-                // erase all the other classes
-                cmd.CommandText = "DELETE FROM Classes" +
-                " WHERE idClass<>" + Class1.IdClass +
-                " AND idClass<>" + Class2.IdClass + ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the lessons of other classes
-                cmd.CommandText = "DELETE FROM Lessons" +
-                    " WHERE idClass<>" + Class1.IdClass +
-                    " AND idClass<>" + Class2.IdClass + ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the students of other classes from the link table
-                cmd.CommandText = "DELETE FROM Classes_Students" +
-                 " WHERE idClass<>" + Class1.IdClass +
-                 " AND idClass<>" + Class2.IdClass + ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the students of other classes 
-                cmd.CommandText = "DELETE FROM Students" +
-                    " WHERE idStudent NOT IN" +
-                    " (SELECT idStudent FROM Classes_Students" +
-                    " WHERE idClass<>" + Class1.IdClass +
-                    " OR idClass<>" + Class2.IdClass +
-                    ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the annotation, of all classes
-                cmd.CommandText = "DELETE FROM StudentsAnnotations" +
-                    ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the StartLinks of other classes
-                cmd.CommandText = "DELETE FROM Classes_StartLinks" +
-                    " WHERE idClass<>" + Class1.IdClass +
-                    " AND idClass<>" + Class2.IdClass +
-                    ";";
-                cmd.ExecuteNonQuery();
-
-                // erase all the grades of other classes' students
-                cmd.CommandText = "DELETE FROM Grades" +
-                    " WHERE idStudent NOT IN" +
-                    " (SELECT idStudent FROM Classes_Students" +
-                    " WHERE idClass<>" + Class1.IdClass +
-                    " OR idClass<>" + Class2.IdClass +
-                    ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the links to photos of other classes' students
-                cmd.CommandText = "DELETE FROM StudentsPhotos_Students" +
-                    " WHERE idStudent NOT IN" +
-                    " (SELECT idStudent FROM Classes_Students" +
-                    " WHERE idClass<>" + Class1.IdClass +
-                    " OR idClass<>" + Class2.IdClass +
-                    ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the photos of other classes' students
-                cmd.CommandText = "DELETE FROM StudentsPhotos WHERE StudentsPhotos.idStudentsPhoto NOT IN" +
-                    "(SELECT StudentsPhotos_Students.idStudentsPhoto" +
-                    " FROM StudentsPhotos, StudentsPhotos_Students, Classes_Students" +
-                    " WHERE StudentsPhotos_Students.idStudent = Classes_Students.idStudent" +
-                    " AND StudentsPhotos.idStudentsPhoto = StudentsPhotos_Students.idStudentsPhoto" +
-                    " AND (Classes_Students.idClass=" + Class1.IdClass +
-                    " OR Classes_Students.idClass=" + Class2.IdClass + ")" +
-                    ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the images of other classes
-                cmd.CommandText = "DELETE FROM Images WHERE Images.idImage NOT IN" +
-                    "(SELECT DISTINCT Lessons_Images.idImage" +
-                    " FROM Images, Lessons_Images, Lessons" +
-                    " WHERE Lessons_Images.idImage = Images.idImage" +
-                    " AND Lessons_Images.idLesson = Lessons.idLesson" +
-                    " AND (Lessons.idClass=" + Class1.IdClass +
-                    " OR Lessons.idClass=" + Class2.IdClass + ")" +
-                    ");";
-                cmd.ExecuteNonQuery();
-
-                //erase all links to the images of other classes
-                cmd.CommandText = "DELETE FROM Lessons_Images WHERE Lessons_Images.idImage NOT IN" +
-                    "(SELECT DISTINCT Lessons_Images.idImage" +
-                    " FROM Images, Lessons_Images, Lessons" +
-                    " WHERE Lessons_Images.idImage = Images.idImage" +
-                    " AND Lessons_Images.idLesson = Lessons.idLesson" +
-                    " AND (Lessons.idClass=" + Class1.IdClass +
-                    " OR Lessons.idClass=" + Class2.IdClass + ")" +
-                    ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the questions of the students of the other classes
-                // !! StudentsQuestions currently not used !!
-                cmd.CommandText = "DELETE FROM StudentsQuestions" +
-                    " WHERE idStudent NOT IN" +
-                    " (SELECT DISTINCT idStudent FROM Classes_Students" +
-                    " WHERE idClass=" + Class1.IdClass +
-                    " OR idClass=" + Class2.IdClass +
-                    ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the answers  of the students of the other classes
-                // !! StudentsAnswers currently not used !!
-                cmd.CommandText = "DELETE FROM StudentsAnswers" +
-                " WHERE idStudent NOT IN" +
-                " (SELECT idStudent FROM Classes_Students" +
-                    " WHERE idClass=" + Class1.IdClass +
-                    " OR idClass=" + Class2.IdClass + ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the tests of students of the other classes
-                // !! StudentsTests currently not used !!
-                cmd.CommandText = "DELETE FROM StudentsTests" +
-                " WHERE idStudent NOT IN" +
-                " (SELECT idStudent FROM Classes_Students" +
-                " WHERE idClass=" + Class1.IdClass +
-                    " OR idClass=" + Class2.IdClass +
-                ");";
-                cmd.ExecuteNonQuery();
-
-                // erase all the topics of other classes' lessons
-                cmd.CommandText = "DELETE FROM Lessons_Topics" +
-                    " WHERE idLesson NOT IN" +
-                    " (SELECT idLesson from Lessons" +
-                    " WHERE idClass=" + Class1.IdClass +
-                    " OR idClass=" + Class2.IdClass +
-                    ");";
-                cmd.ExecuteNonQuery();
-
-                // change the data of the classes
-                Class1.Abbreviation = "DEMO1";
-                Class1.Description = "SchoolGrades demo class 1";
-                // Class1.IdSchool = ""; // left the existing code 
-                Class1.PathRestrictedApplication = Commons.PathExe + "\\demo1";
-                // Class1.SchoolYear = // !!!! shift the data to the destination school year, to be done when year's shifting will be managed!!!!
-                Class1.IdSchool = Commons.IdSchool;
-                Class1.UriWebApp = ""; // ???? decide what to put here ????
-
-                // SaveClass Class1;
-                string query = "UPDATE Classes" +
-                    " SET" +
-                    " idClass=" + Class1.IdClass + "" +
-                    ",idSchoolYear='" + SqlVal.SqlString(Class1.SchoolYear) + "'" +
-                    ",idSchool='" + SqlVal.SqlString(Class1.IdSchool) + "'" +
-                    ",abbreviation='" + SqlVal.SqlString(Class1.Abbreviation) + "'" +
-                    ",desc='" + SqlVal.SqlString(Class1.Description) + "'" +
-                    ",uriWebApp='" + Class1.UriWebApp + "'" +
-                    ",pathRestrictedApplication='" + SqlVal.SqlString(Class1.PathRestrictedApplication) + "'" +
-                    " WHERE idClass=" + Class1.IdClass +
-                    ";";
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-
-                Class2.Abbreviation = "DEMO2";
-                Class2.Description = "SchoolGrades demo class 2";
-                Class2.PathRestrictedApplication = Commons.PathExe + "\\demo2";
-                // Class2.SchoolYear = !!!! shift the data to the destination school year !!!!
-                Class2.IdSchool = Commons.IdSchool;
-                Class2.UriWebApp = ""; // ???? decide what to put here ????
-                // SaveClass Class2;
-                query = "UPDATE Classes" +
-                    " SET" +
-                    " idClass=" + Class2.IdClass + "" +
-                    ",idSchoolYear='" + SqlVal.SqlString(Class2.SchoolYear) + "'" +
-                    ",idSchool='" + SqlVal.SqlString(Class2.IdSchool) + "'" +
-                    ",abbreviation='" + SqlVal.SqlString(Class2.Abbreviation) + "'" +
-                    ",desc='" + SqlVal.SqlString(Class2.Description) + "'" +
-                    ",uriWebApp='" + Class2.UriWebApp + "'" +
-                    ",pathRestrictedApplication='" + SqlVal.SqlString(Class2.PathRestrictedApplication) + "'" +
-                    " WHERE idClass=" + Class2.IdClass +
-                    ";";
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-
-                // erase all the users
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Users" +
-                    ";";
-                cmd.ExecuteNonQuery();
-
-                // rename every student left in the database according to the names found in the pictures' filenames
-                RenameStudentsNamesFromPictures(Class1, conn);
-                RenameStudentsNamesFromPictures(Class2, conn);
-
-                // change the paths of the images 
-                ChangeImagesPath(Class1, conn);
-                ChangeImagesPath(Class2, conn);
-
-                // randomly change all grades 
-                dl.RandomizeGrades(conn);
-
-                // change the lesson dates to this school year (when we implement year shift!) 
-                // !!!! TODO !!!!
-
-                // change the school year in StudentsPhotos_Students (when we implement year shift!) 
-                // !!!! TODO !!!!
-
-                // compact the database 
-                cmd.CommandText = "VACUUM;";
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-            return newDatabaseFullName;
-        }
-        private void ChangeImagesPath(Class Class, DbConnection conn)
-        {
-            DbDataReader dRead;
-            DbCommand cmd = conn.CreateCommand();
-            cmd = conn.CreateCommand();
-            cmd.CommandText = "SELECT Images.idImage, Images.imagePath" +
-                " FROM Images" +
-                " JOIN Lessons_Images ON Images.idImage=Lessons_Images.idImage" +
-                " JOIN Lessons ON Lessons.idLesson = Lessons_Images.idLesson" +
-                " WHERE Lessons.idClass=" + Class.IdClass +
-            ";";
-            dRead = cmd.ExecuteReader();
-            string newFolder = Class.SchoolYear + Class.Abbreviation;
-            while (dRead.Read())
-            {
-                string path = SafeDb.SafeString(dRead["imagePath"]);
-                int? id = SafeDb.SafeInt(dRead["idImage"]);
-                string partToReplace = path.Substring(0, path.IndexOf("\\"));
-                path = path.Replace(partToReplace, newFolder);
-                SaveImagePath(id, path, conn);
-            }
-            cmd.Dispose();
-        }
-
-        private void SaveImagePath(int? id, string path, DbConnection conn)
-        {
-            DbCommand cmd = conn.CreateCommand();
-            cmd.CommandText = "UPDATE Images" +
-            " SET imagePath='" + SqlVal.SqlString(path) + "'" +
-            " WHERE idImage=" + id +
-            ";";
-            cmd.ExecuteNonQuery();
-            cmd.Dispose();
-        }
-        private void SaveStudentsPhotosPath(int? id, string path, DbConnection conn)
-        {
-            if (id != null)
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "UPDATE StudentsPhotos" +
-                " SET photoPath='" + SqlVal.SqlString(path) + "'" +
-                " WHERE idStudentsPhoto=" + id +
-                ";";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-
-        private void RenameStudentsNamesFromPictures(Class Class, DbConnection conn)
-        {
-            // get the "previous" students from database 
-            List<Student> StudentsInClass = dl.GetStudentsOfClass(Class.IdClass, conn);
-
-            // rename the students' names according to the names found in the image files 
-            string[] OriginalDemoPictures = Directory.GetFiles(Commons.PathImages + "\\DemoPictures\\");
-            // start assigning the names from a random image
-            Random rnd = new Random();
-            int pictureIndex = rnd.Next(0, OriginalDemoPictures.Length - 1);
-            foreach (Student s in StudentsInClass)
-            {
-                string justFileName = Path.GetFileName(OriginalDemoPictures[pictureIndex]);
-                string fileWithNoExtension = justFileName.Substring(0, justFileName.LastIndexOf('.'));
-                string[] wordsInFileName = (Path.GetFileName(fileWithNoExtension)).Split(' ');
-                string lastName = "";
-                string firstName = "";
-                foreach (string word in wordsInFileName)
-                {
-                    if (word == word.ToUpper())
-                    {
-                        lastName += " " + word;
-                    }
-                    else
-                    {
-                        firstName += " " + word;
-                    }
-                }
-                lastName = lastName.Trim();
-                firstName = firstName.Trim();
-
-                s.LastName = lastName;
-                s.FirstName = firstName;
-                s.BirthDate = null;
-                s.BirthPlace = null;
-                s.Class = "";
-                s.Email = "";
-                s.IdClass = 0;
-                s.ArithmeticMean = 0;
-                s.RegisterNumber = null;
-                s.Residence = null;
-                s.RevengeFactorCounter = 0;
-                s.Origin = null;
-                s.SchoolYear = null;
-                s.Sum = 0;
-                dl.SaveStudent(s, conn);
-
-                // save the image with standard name in the folder of the demo class
-                string fileExtension = Path.GetExtension(OriginalDemoPictures[pictureIndex]);
-                string folder = Commons.PathImages + "\\" + Class.SchoolYear + Class.Abbreviation + "\\";
-                string filename = s.LastName + "_" + s.FirstName + "_" + Class.Abbreviation + Class.SchoolYear + fileExtension;
-                if (!Directory.Exists(folder))
-                {
-                    Directory.CreateDirectory(folder);
-                }
-                if (File.Exists(folder + filename))
-                {
-                    File.Delete(folder + filename);
-                }
-                File.Copy(OriginalDemoPictures[pictureIndex], folder + filename);
-
-                // change student pictures' paths in table StudentsPhotos
-                string relativePathAndFile = Class.SchoolYear + Class.Abbreviation + "\\" + filename;
-                int? idImage = GetStudentsPhotoId(s.IdStudent, Class.SchoolYear, conn);
-                SaveStudentsPhotosPath(idImage, relativePathAndFile, conn);
-
-                // copy all the lessons images files that aren't already there or that have a newer date 
-                string query = "SELECT Images.imagePath, Classes.pathRestrictedApplication" +
-                " FROM Images" +
-                    " JOIN Lessons_Images ON Lessons_Images.idImage=Images.idImage" +
-                    " JOIN Lessons ON Lessons_Images.idLesson=Lessons.idLesson" +
-                    " JOIN Classes ON Classes.idClass=Lessons.idClass" +
-                    " WHERE Lessons.idClass=" + Class.IdClass +
-                    ";";
-                DbCommand cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                DbDataReader dReader = cmd.ExecuteReader();
-                while (dReader.Read())
-                {
-                    //string destinationFile = (string)dReader["pathRestrictedApplication"] +
-                    //    "\\SchoolGrades\\" + "Images" + "\\" + (string)dReader["imagePath"];
-                    string filePart = (string)dReader["imagePath"];
-                    string partToReplace = filePart.Substring(0, filePart.IndexOf("\\"));
-                    filePart = filePart.Replace(partToReplace, Class.SchoolYear + Class.Abbreviation);
-                    string destinationFile = (string)Commons.PathImages + "\\" + filePart;
-
-                    if (!Directory.Exists(Path.GetDirectoryName(destinationFile)))
-                    {
-                        Directory.CreateDirectory(Path.GetDirectoryName(destinationFile));
-                    }
-                    if (!File.Exists(destinationFile) ||
-                        File.GetLastWriteTime(destinationFile)
-                        < File.GetLastWriteTime(Commons.PathImages + "\\" + (string)dReader["imagePath"]))
-                        // destination file not existing or older
-                        try
-                        {
-                            File.Copy(Commons.PathImages + "\\" + (string)dReader["imagePath"],
-                                destinationFile);
-                        }
-                        catch (Exception ex)
-                        {
-                            Console.Beep();
-                        }
-                }
-                dReader.Dispose();
-
-                if (++pictureIndex >= OriginalDemoPictures.Length)
-                    pictureIndex = 0;
-            }
-        }
-
-        private Nullable<int> GetStudentsPhotoId(int? idStudent, string schoolYear, DbConnection conn)
-        {
-            DbCommand cmd = conn.CreateCommand();
-            cmd.CommandText = "SELECT idStudentsPhoto FROM StudentsPhotos_Students " +
-                "WHERE idStudent=" + idStudent + " AND idSchoolYear=" + schoolYear + "" +
-                ";";
-            return (int?)cmd.ExecuteScalar();
-        }
-
-        
-
-       
-        internal int CreateNewTopic(Topic NewTopic)
-        {
-            int nextId;
-            using (DbConnection conn = dl.Connect())
-            {
-                nextId = dl.NextKey("Topics", "idTopic");
-
-                DbCommand cmd = conn.CreateCommand();
-                // aggiunge la foto alle foto (cartella relativa, cui verrà aggiunta la path delle foto)
-                cmd.CommandText = "INSERT INTO Topics " +
-                "(idTopic,name,desc,leftNode,rightNode,parentNode,childNumber)" +
-                "Values " +
-                "(" + nextId + ",'" + SqlVal.SqlString(NewTopic.Name) + "','" +
-                SqlVal.SqlString(NewTopic.Desc) + "'," +  SqlVal.SqlInt(NewTopic.LeftNodeNew.ToString()) + "," +
-                 SqlVal.SqlInt(NewTopic.RightNodeNew.ToString()) + "," +  SqlVal.SqlInt(NewTopic.ParentNodeNew.ToString()) +
-                "," + SqlVal.SqlInt(NewTopic.ChildNumberNew.ToString()) +
-                ");";
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-            return nextId;
-        }
-        internal void EraseAllTopics()
-        {
-            using (DbConnection conn = dl.Connect())
-            {   // erase all the topics
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Topics;";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-        /// <summary>
-        /// Gets the record of the Topic from the database, 
-        /// </summary>
-        /// <param name="dRead"></param>
-        /// <returns></returns>
-        internal Topic GetTopicFromRow(DbDataReader dRead)
-        {
-            Topic t = new Topic();
-            t.Id = SafeDb.SafeInt(dRead["IdTopic"]);
-            t.Name = SafeDb.SafeString(dRead["name"]);
-            t.Desc = SafeDb.SafeString(dRead["desc"]);
-            t.LeftNodeOld = SafeDb.SafeInt(dRead["leftNode"]);
-            t.LeftNodeNew = -1;
-            t.RightNodeOld = SafeDb.SafeInt(dRead["rightNode"]);
-            t.RightNodeNew = -1;
-            t.ParentNodeOld = SafeDb.SafeInt(dRead["parentNode"]);
-            t.ParentNodeNew = -1;
-            t.ChildNumberOld = SafeDb.SafeInt(dRead["childNumber"]);
-            t.ChildNumberNew = -1;
-            t.Changed = false;
-
-            return t;
-        }
-        internal Topic GetTopicById(int? idTopic)
-        {
-            Topic t = new Topic();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT *" +
-                    " FROM Topics" +
-                    " WHERE idTopic=" + idTopic;
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                DbDataReader dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    t = GetTopicFromRow(dRead);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return t;
-        }
-        internal List<Topic> GetTopics()
-        {
-            List<Topic> lt = new List<Topic>();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT *" +
-                    " FROM Topics" +
-                    " ORDER BY IdTopic;";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                DbDataReader dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Topic t = GetTopicFromRow(dRead);
-                    lt.Add(t);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return lt;
-        }
-        internal List<Topic> GetTopicsNotDoneFromThisTopic(Class Class, Topic Topic,
-            SchoolSubject Subject)
-        {
-            // node numbering according to Modified Preorder Tree Traversal algorithm
-            List<Topic> l = new List<Topic>();
-            using (DbConnection conn = dl.Connect())
-            {
-                // find descendant topics that aren't done  
-                DbCommand cmd = conn.CreateCommand();
-                //!!!! TODO aggiustare la logica
-                string query = "SELECT *" +
-                    " FROM Topics" +
-                    " LEFT JOIN Lessons_Topics ON Lessons_Topics.idTopic = Topics.idTopic" +
-                    " JOIN Lessons ON Lessons_Topics.idLesson = Lessons.IdLesson" +
-                    " WHERE leftNode BETWEEN " + Topic.LeftNodeOld +
-                    " AND " + Topic.RightNodeOld +
-                    " AND Lessons_Topics.idLesson IS null" +
-                    " AND Lessons.idClass = " + Class.IdClass +
-                    " AND Lessons.idSchoolSubject ='" + Subject.IdSchoolSubject + "'" +
-                    " ORDER BY leftNode ASC;";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                DbDataReader dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Topic t = GetTopicFromRow(dRead);
-                    l.Add(t);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return l;
-        }
-        internal List<Topic> GetTopicsDoneFromThisTopic(Class Class, Topic StartTopic,
-            SchoolSubject Subject)
-        {
-            // node order according to Modified Preorder Tree Traversal algorithm
-            List<Topic> l = new List<Topic>();
-            if (Class == null)
-                return l;
-            using (DbConnection conn = dl.Connect())
-            {
-                // find descendant topics that are done  
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT *" +
-                    " FROM Topics" +
-                    " JOIN Lessons_Topics ON Lessons_Topics.idTopic = Topics.idTopic " +
-                    " JOIN Lessons ON Lessons_Topics.idLesson = Lessons.idLesson" +
-                    " WHERE leftNode BETWEEN " + StartTopic.LeftNodeOld +
-                    " AND " + StartTopic.RightNodeOld;
-                if (Class != null)
-                    query += " AND Lessons.idClass = " + Class.IdClass; 
-                if(Subject != null)
-                    query += " AND Lessons.idSchoolSubject ='" + Subject.IdSchoolSubject + "'" +
-                    " ORDER BY leftNode ASC;";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                DbDataReader dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Topic t = GetTopicFromRow(dRead);
-                    l.Add(t);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return l;
-        }
-        internal List<Topic> GetAllTopicsDoneInClassAndSubject(Class Class,
-            SchoolSubject Subject, 
-            DateTime DateStart = default(DateTime), DateTime DateFinish = default(DateTime))
-        {
-            // node order according to Modified Preorder Tree Traversal algorithm
-            List<Topic> l = new List<Topic>();
-            using (DbConnection conn = dl.Connect())
-            {
-                // find topics that are done in a lesson of given class about and given subject 
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT *" +
-                    " FROM Topics" +
-                    " JOIN Lessons_Topics ON Lessons_Topics.idTopic = Topics.idTopic " +
-                    " JOIN Lessons ON Lessons_Topics.idLesson = Lessons.idLesson" +
-                    " JOIN Classes ON Classes.idClass = Lessons.idClass" +
-                    " WHERE Lessons.idClass = " + Class.IdClass +
-                    " AND Lessons.idSchoolSubject ='" + Subject.IdSchoolSubject + "'";
-                if (DateStart != default(DateTime) && DateFinish != default(DateTime))
-                    query += " AND Lessons.date BETWEEN " +
-                    SqlVal.SqlDate(DateStart) + " AND " + SqlVal.SqlDate(DateFinish); 
-                    query += " ORDER BY Lessons.date ASC;";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                DbDataReader dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Topic t = GetTopicFromRow(dRead);
-                    l.Add(t);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return l;
-        }
-        internal List<Topic> GetTopicsDoneInPeriod(Class currentClass, SchoolSubject currentSubject,
-            DateTime DateFrom, DateTime DateTo)
-        {
-            List<Topic> lt = new List<Topic>();
-            using (DbConnection conn = dl.Connect())
-            {
-                DateTo = DateTo.AddDays(1); // add one day for lesson after 0 and to midnight 
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT Topics.idTopic,Topics.name,Topics.desc,Topics.LeftNode,Topics.RightNode," +
-                    "Topics.ParentNode, Topics.childNumber, Lessons.date,Lessons.idSchoolSubject" +
-                    " FROM Topics" +
-                    " JOIN Lessons_Topics ON Lessons_Topics.IdTopic=Topics.IdTopic" +
-                    " JOIN Lessons ON Lessons_Topics.IdLesson=Lessons.IdLesson" +
-                    " WHERE Lessons.IdClass=" + currentClass.IdClass +
-                    " AND Lessons.idSchoolSubject='" + currentSubject.IdSchoolSubject + "'";
-                if (DateFrom == Commons.DateNull)
-                {
-                    query += " AND (Lessons.Date BETWEEN '" + DateFrom.ToString("yyyy-MM-dd") + "'" +
-                        " AND '" + DateTo.ToString("yyyy-MM-dd") + "')";
-                }
-                query += " ORDER BY Lessons.date DESC" +
-                ";";
-                cmd = new SQLiteCommand(query);
-                cmd.Connection = conn;
-                DbDataReader dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Topic t = GetTopicFromRow(dRead);
-                    t.Id = (int)dRead["IdTopic"];
-                    t.Name = (string)dRead["name"];
-                    t.Desc = (string)dRead["desc"];
-                    //t.LeftNodeNew = -1;
-                    //t.RightNodeNew = -1;
-                    t.Date = (DateTime)dRead["date"]; // taken fron the Lessons table 
-
-                    // determine the path while still in the database
-                    // if we don't, determination from the outside would be too costly 
-                    query = "SELECT idTopic, name, desc, leftNode, rightNode" +
-                        " FROM Topics" +
-                        " WHERE leftNode <=" + t.LeftNodeOld +
-                        " AND rightNode >=" + t.RightNodeOld +
-                        " ORDER BY leftNode ASC;)";
-                    cmd = new SQLiteCommand(query);
-                    cmd.Connection = conn;
-                    DbDataReader dRead1 = cmd.ExecuteReader();
-                    string path = "";
-                    while (dRead1.Read())
-                    {
-                        path += ((string)dRead1["name"]).Trim() + "|";
-                    }
-                    //t.Path = path;
-                    t.Changed = false;
-                    lt.Add(t);
-                    dRead1.Dispose();
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-            }
-            return lt;
-        }
-        internal int GetTopicDescendantsNumber(int? LeftNode, int? RightNode)
-        {
-            // node numbering according to Modified Preorder Tree Traversal algorithm
-            return ((int)RightNode - (int)LeftNode - 1) / 2;
-        }
-        internal void UpdateTopic(Topic t, DbConnection conn)
-        {
-            bool leaveConnectionOpen = true;
-            if (conn == null)
-            {
-                conn = dl.Connect();
-                leaveConnectionOpen = false;
-            }
-            DbCommand cmd = conn.CreateCommand();
-            cmd.CommandText = "UPDATE Topics" +
-                " SET" +
-                " name='" + SqlVal.SqlString(t.Name) + "'" +
-                ",desc='" + SqlVal.SqlString(t.Desc) + "'" +
-                ",parentNode=" + t.ParentNodeNew +
-                ",leftNode=" + t.LeftNodeNew +
-                ",rightNode=" + t.RightNodeNew +
-                ",childNumber=" + t.ChildNumberNew +
-                " WHERE idTopic=" + t.Id +
-                ";";
-            cmd.ExecuteNonQuery();
-            cmd.Dispose();
-            if (!leaveConnectionOpen)
-            {
-                conn.Close();
-                conn.Dispose();
-            }
-        }
-        internal void InsertTopic(Topic t, DbConnection Conn)
-        {
-            if (t.Id == null || t.Id == 0)
-            {
-                bool leaveConnectionOpen = true;
-                if (Conn == null)
-                {
-                    Conn = dl.Connect();
-                    leaveConnectionOpen = false;
-                }
-                DbCommand cmd = Conn.CreateCommand();
-
-                cmd.CommandText = "SELECT MAX(IdTopic) FROM Topics;";
-                var temp = cmd.ExecuteScalar();
-                if (!(temp is DBNull))
-                    t.Id = Convert.ToInt32(temp) + 1;
-                cmd.CommandText = "INSERT INTO Topics" +
-                    " (idTopic,name,desc,leftNode,rightNode,parentNode,childNumber)" +
-                    " Values (" +
-                    t.Id.ToString() +
-                    ",'" + SqlVal.SqlString(t.Name) + "'" +
-                    ",'" + SqlVal.SqlString(t.Desc) + "'" +
-                    "," + t.LeftNodeNew + "" +
-                    "," + t.RightNodeNew + "" +
-                    "," + t.ParentNodeNew + "" +
-                    "," + t.ChildNumberNew + "" +
-                    ");";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-                if (!leaveConnectionOpen)
-                {
-                    Conn.Close();
-                    Conn.Dispose();
-                }
-            }
-        }
-        internal void SaveTopicsFromScratch(List<Topic> ListTopics)
-        {
-            ////////BackgroundCanStillSaveTopicsTree = true;
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "DELETE FROM Topics;";
-                cmd.ExecuteNonQuery();
-                int key;
-                cmd.CommandText = "SELECT MAX(IdTopic) FROM Topics;";
-                var temp = cmd.ExecuteScalar();
-                if (temp is DBNull)
-                    key = 0;
-                else
-                    key = (int)temp;
-                foreach (Topic t in ListTopics)
-                {   // insert new nodes
-                    {
-                        cmd.CommandText = "INSERT INTO Topics" +
-                           " (idTopic,name,desc,parentNode,leftNode,rightNode,parentNode)" +
-                           " Values (" +
-                           (++key).ToString() +
-                            ",'" + SqlVal.SqlString(t.Name) + "'" +
-                            ",'" + SqlVal.SqlString(t.Desc) + "'" +
-                            "," + t.ParentNodeNew + "" +
-                            "," + t.LeftNodeNew + "" +
-                            "," + t.RightNodeNew + "" +
-                            "," + t.ParentNodeNew + "" +
-                            ");";
-                        cmd.ExecuteNonQuery();
-                    }
-                }
-                cmd.Dispose();
-            }
-        }
-
-        internal List<QuestionType> GetListQuestionTypes(bool IncludeANullObject)
-        {
-            List<QuestionType> l = new List<QuestionType>();
-            if (IncludeANullObject)
-            {
-                QuestionType qt = new QuestionType();
-                qt.IdQuestionType = "";
-                l.Add(qt);
-            }
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT * FROM QuestionTypes;";
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    QuestionType o = new QuestionType();
-                    o.Name = (string)dRead["Name"];
-                    o.IdQuestionType = (string)dRead["IdQuestionType"];
-                    o.Desc = (string)dRead["Desc"];
-                    o.IdQuestionType = (string)dRead["IdQuestionType"];
-                    l.Add(o);
-                }
-                dRead.Dispose();
-                cmd.Dispose();
-                return l;
-            }
-        }
-
-        internal int NewLesson(Lesson Lesson)
-        {
-            int key;
-            using (DbConnection conn = dl.Connect())
-            {
-                key = dl.NextKey("Lessons", "idLesson");
-                Lesson.IdLesson = key;
-                // add new record to Lessons table
-                DbCommand cmd = conn.CreateCommand();
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "INSERT INTO Lessons" +
-                " (idLesson, date, idClass, idSchoolSubject, idSchoolYear, note) " +
-                "Values (" +
-                "" + Lesson.IdLesson + "" +
-                "," + SqlVal.SqlDate(Lesson.Date) + "" +
-                "," + Lesson.IdClass + "" +
-                ",'" + Lesson.IdSchoolSubject + "'" +
-                ",'" + Lesson.IdSchoolYear + "'" +
-                ",'" + SqlVal.SqlString(Lesson.Note) + "'" +
-                ");";
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-            return key;
-        }
-        internal void SaveLesson(Lesson Lesson)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd = conn.CreateCommand();
-                cmd.CommandText = "UPDATE Lessons" +
-                " SET" +
-                " date=" + SqlVal.SqlDate(Lesson.Date) + "," +
-                " idClass=" + Lesson.IdClass + "," +
-                " idSchoolSubject='" + Lesson.IdSchoolSubject + "'," +
-                " idSchoolYear='" + Lesson.IdSchoolYear + "'," +
-                " note='" + SqlVal.SqlString(Lesson.Note) + "'" +
-                " WHERE idLesson=" + Lesson.IdLesson +
-                ";";
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-        }
-        internal object GetTopicsOfOneLessonOfClass(Class Class, Lesson Lesson)
-        {
-            DataTable t;
-            using (DbConnection conn = dl.Connect())
-            {
-                DataAdapter dAdapt;
-                DataSet dSet = new DataSet();
-                string query = "SELECT Topics.* FROM Lessons" +
-                    " LEFT JOIN Lessons_Topics ON Lessons.IdLesson = Lessons_Topics.idLesson" +
-                    " LEFT JOIN Topics ON Topics.idTopic = Lessons_Topics.idTopic" +
-                    " WHERE idSchoolSubject='" + Lesson.IdSchoolSubject + "'" +
-                    " AND Lessons.idSchoolYear='" + Lesson.IdSchoolYear + "'" +
-                    " AND Lessons.idClass='" + Class.IdClass + "'" +
-                    " AND Lessons.idLesson='" + Lesson.IdLesson + "'" +
-                    //" GROUP BY Lessons.idLesson" +
-                    " ORDER BY Lessons.date DESC" +
-                    ";";
-                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                dSet = new DataSet("GetOnLessonOfClass");
-                dAdapt.Fill(dSet);
-                t = dSet.Tables[0];
-
-                dAdapt.Dispose();
-                dSet.Dispose();
-            }
-            return t;
-        }
-        internal DataTable GetLessonsOfClass(Class Class, Lesson Lesson)
-        {
-            DataTable t;
-            using (DbConnection conn = dl.Connect())
-            {
-                DataAdapter dAdapt;
-                DataSet dSet = new DataSet();
-                string query = "SELECT * FROM Lessons" +
-                    " WHERE idSchoolSubject='" + Lesson.IdSchoolSubject + "'" +
-                    " AND Lessons.idSchoolYear='" + Lesson.IdSchoolYear + "'" +
-                    " AND Lessons.idClass='" + Class.IdClass + "'" +
-                    //" GROUP BY Lessons.idLesson" +
-                    " ORDER BY Lessons.date DESC" +
-                    ";";
-                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
-                dSet = new DataSet("GetLessonsOfClass");
-                dAdapt.Fill(dSet);
-                t = dSet.Tables[0];
-
-                dAdapt.Dispose();
-                dSet.Dispose();
-            }
-            return t;
-        }
-        internal Lesson GetLastLesson(Lesson CurrentLesson)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                List<Couple> couples = new List<Couple>();
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                Lesson l;
-                string query;
-                query = "SELECT * FROM Lessons" +
-                        " WHERE idClass=" + CurrentLesson.IdClass.ToString() +
-                        " AND idSchoolSubject='" + CurrentLesson.IdSchoolSubject + "'" +
-                        " AND idSchoolYear='" + CurrentLesson.IdSchoolYear + "'" +
-                        " ORDER BY Date DESC LIMIT 1;";
-                cmd.CommandText = query;
-                dRead = cmd.ExecuteReader();
-                l = new Lesson();
-                while (dRead.Read())
-                {
-                    l.IdLesson = (int)dRead["idLesson"];
-                    l.Date = (DateTime)dRead["Date"];
-                    l.IdClass = (int)dRead["idClass"];
-                    l.IdSchoolSubject = (string)dRead["idSchoolSubject"];
-                    l.Note = (string)dRead["note"];
-
-                    break;
-                }
-                cmd.Dispose();
-                dRead.Dispose();
-                return l;
-            }
-        }
-        internal Lesson GetLessonInDate(Class Class, string IdSubject,
-            DateTime Date)
-        {
-            Lesson less = new Lesson();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbDataReader dRead;
-                DbCommand cmd = conn.CreateCommand();
-                string query;
-                query = "SELECT * FROM Lessons" +
-                        " WHERE idClass=" + Class.IdClass.ToString() +
-                        " AND idSchoolSubject='" + IdSubject + "'" +
-                        " AND date BETWEEN " + SqlVal.SqlDate(Date.ToString("yyyy-MM-dd")) +
-                        " AND " + SqlVal.SqlDate(Date.AddDays(1).ToString("yyyy-MM-dd")) +
-                        " LIMIT 1;";
-                cmd.CommandText = query;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    less.IdLesson = (int)dRead["idLesson"];
-                    less.Date = (DateTime)dRead["Date"];
-                    less.IdClass = (int)dRead["idClass"];
-                    less.IdSchoolSubject = (string)dRead["idSchoolSubject"];
-                    less.Note = (string)dRead["note"];
-
-                    break; // there should be only one record in the query result 
-                }
-                cmd.Dispose();
-                dRead.Dispose();
-            }
-            return less;
-        }
-        internal void EraseLesson(int? IdLesson, bool AlsoEraseImageFiles)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                // erase existing links topic-lesson
-                DbCommand cmd = conn.CreateCommand();
-                string query = "DELETE FROM Lessons_Topics" +
-                        " WHERE idLesson=" + IdLesson.ToString() +
-                        ";";
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-
-                // erase images' files if permitted
-                if (AlsoEraseImageFiles)
-                {
-                    // !! TODO !! find the images that aren't linked to another lesson and delete
-                    // the files if EraseImageFiles is set 
-                    throw new NotImplementedException();
-                }
-
-                // erase existing links images-lesson
-                cmd = conn.CreateCommand();
-                query = "DELETE FROM Lessons_Images" +
-                        " WHERE idLesson=" + IdLesson.ToString() +
-                        ";";
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-
-                // erase the lesson's row from lessons
-                cmd = conn.CreateCommand();
-                query = "DELETE FROM Lessons" +
-                        " WHERE idLesson=" + IdLesson.ToString() +
-                        ";";
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-        internal void SaveTopicsOfLesson(int? IdLesson, List<Topic> topicsOfTheLesson)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                // erase existing links topic-lesson
-                DbCommand cmd = conn.CreateCommand();
-                string query = "DELETE FROM Lessons_Topics" +
-                        " WHERE idLesson=" + IdLesson.ToString() +
-                        ";";
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-                int insertionOrder = 1;
-                foreach (Topic t in topicsOfTheLesson)
-                {
-                    // insert links topic-lesson, one at a time 
-                    cmd.CommandText = "INSERT INTO Lessons_Topics" +
-                    " (idLesson, idTopic, insertionOrder)" +
-                    " Values (" +
-                    "" + IdLesson + "" +
-                    "," + t.Id +
-                    "," + insertionOrder++ +
-                    ");";
-                    try
-                    {
-                        cmd.ExecuteNonQuery();
-                    }
-
-                    catch (Exception e)
-                    {
-                        {
-                            // if it isn't UNIQUE, then it is error (to cure!!!!)
-                            // but actually it doesn't matter too much, because
-                            // the lesson is already linked to the topic anyway.. 
-
-
-                        }
-                    }
-                }
-                cmd.Dispose();
-            }
-        }
-        internal List<Topic> GetTopicsOfLesson(int? IdLesson, List<Topic> topicsOfTheLesson)
-        {
-            if (IdLesson == null)
-            {
-                return null; 
-            }
-            // order by ensures that the order of the result is the order of insertion 
-            // in the database (that was the same of the tree traversal) 
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                DbDataReader dRead;
-                string query;
-                query = "SELECT * FROM Topics" +
-                        " JOIN Lessons_Topics ON Topics.idTopic=Lessons_Topics.idTopic" +
-                        " WHERE Lessons_Topics.idLesson=" + IdLesson +
-                        " ORDER BY insertionOrder" +
-                        ";";
-                cmd.CommandText = query;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Topic t = GetTopicFromRow(dRead);
-                    topicsOfTheLesson.Add(t);
-                }
-                cmd.Dispose();
-                dRead.Dispose();
-            }
-            return topicsOfTheLesson;
-        }
-        internal List<Image> GetLessonsImagesList(Lesson Lesson)
-        {
-            if (Lesson.IdLesson == null)
-                return null;
-
-            List<Image> imagesOfTheLesson = new List<Image>();
-
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                DbDataReader dRead;
-                string query;
-                query = "SELECT * FROM Images" +
-                        " JOIN Lessons_Images ON Images.idImage=Lessons_Images.idImage" +
-                        " WHERE Lessons_Images.idLesson=" + Lesson.IdLesson +
-                        ";";
-                cmd.CommandText = query;
-                dRead = cmd.ExecuteReader();
-                while (dRead.Read())
-                {
-                    Image i = new Image();
-                    i.IdImage = (int)dRead["IdImage"];
-                    i.Caption = (string)dRead["Caption"];
-                    i.RelativePathAndFilename = (string)dRead["ImagePath"];
-
-                    imagesOfTheLesson.Add(i);
-                }
-                cmd.Dispose();
-                dRead.Dispose();
-            }
-            return imagesOfTheLesson;
-        }
-        /// <summary>
-        /// Creates a new Image in Images and links it to the lesson
-        /// If the image has an id != 0, it exists and is not created 
-        /// </summary>
-        /// <param name="Image"></param>
-        /// <param name="Lesson"></param>
-        /// <returns></returns>
-        internal int? LinkOneImage(Image Image, Lesson Lesson)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query;
-                if (Image.IdImage == 0)
-                {
-                    Image.IdImage = dl.NextKey("Images", "IdImage");
-                    query = "INSERT INTO Images" +
-                    " (idImage, imagePath, caption)" +
-                    " Values (" + Image.IdImage + ",'" +
-                    SqlVal.SqlString(Image.RelativePathAndFilename) + "','" +
-                    SqlVal.SqlString(Image.Caption) + "'" +
-                    ");";
-                    cmd.CommandText = query;
-                    cmd.ExecuteNonQuery();
-                }
-                query = "INSERT INTO Lessons_Images" +
-                    " (idImage, idLesson)" +
-                    " Values (" + Image.IdImage + "," + Lesson.IdLesson + "" +
-                    ");";
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-            return Image.IdImage;
-        }
-
-        internal void RemoveImageFromLesson(Lesson Lesson, Image Image,
-            bool AlsoEraseImageFile)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query;
-                query = "DELETE FROM Lessons_Images" +
-                    " WHERE idImage=" +
-                    Image.IdImage +
-                    ";";
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-
-                if (AlsoEraseImageFile)
-                {
-
-                    query = "DELETE FROM Images" +
-                        " WHERE idImage=" +
-                        Image.IdImage +
-                        ";";
-                    cmd.CommandText = query;
-                    cmd.ExecuteNonQuery();
-                    try
-                    {
-                        File.Delete(Commons.PathImages + "\\" + Image.RelativePathAndFilename);
-                    }
-                    catch (Exception ex)
-                    {
-                        string err = "DbLayer|RemoveImageFromLesson|" +
-                            Commons.PathImages + "\\" + Image.RelativePathAndFilename +
-                            ".\r\n" + ex.Message + ex.StackTrace;
-                        Commons.ErrorLog(err);
-                        throw new Exception(err);
-                    }
-                }
-                cmd.Dispose();
-            }
-        }
-
-        internal void SaveImage(Image Image)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query;
-                query = "UPDATE Images" +
-                    " SET caption='" + SqlVal.SqlString(Image.Caption) + "'" +
-                    " WHERE idImage=" +
-                    Image.IdImage +
-                    ";";
-                cmd.CommandText = query;
-                cmd.ExecuteNonQuery();
-
-                cmd.Dispose();
-            }
-        }
-
-        internal Image FindImageWithGivenFile(string PathAndFileNameOfImage)
-        {
-            Image i = new Image();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                DbDataReader dRead;
-                string query;
-                query = "SELECT * FROM Images" +
-                        " WHERE Images.imagePath='" +
-                        SqlVal.SqlString(PathAndFileNameOfImage.Remove(0, Commons.PathImages.Length + 1)) +
-                        "';";
-                cmd.CommandText = query;
-                dRead = cmd.ExecuteReader();
-                dRead.Read(); // just one record ! 
-                if (!dRead.HasRows)
-                    return null;
-                i.IdImage = (int)dRead["IdImage"];
-                i.Caption = (string)dRead["Caption"];
-                i.RelativePathAndFilename = (string)dRead["ImagePath"];
-                cmd.Dispose();
-                dRead.Dispose();
-            }
-            return i;
-        }
+      
         internal List<SchoolPeriod> GetSchoolPeriodsOfDate(DateTime Date)
         {
             List<SchoolPeriod> l = new List<SchoolPeriod>();
@@ -2943,151 +997,6 @@ namespace SchoolGrades.DbClasses
                 cmd.Dispose();
             }
         }
-        internal List<Question> GetAllQuestionsOfATest(int? IdTest)
-        {
-            List<Question> lq = new List<Question>();
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "SELECT *" +
-                    " FROM Questions" +
-                    " JOIN Tests_Questions ON Tests_Questions.IdQuestion=Questions.IdQuestion" +
-                    " WHERE Tests_Questions.idTest=" + IdTest +
-                    ";";
-                DbDataReader dRead;
-                dRead = cmd.ExecuteReader();
-
-                while (dRead.Read())
-                {
-                    Question q = GetQuestionFromRow(dRead);
-                    lq.Add(q);
-                }
-            }
-            return lq;
-        }
-        internal void AddQuestionToTest(Test Test, Question Question)
-        {
-            using (DbConnection conn = dl.Connect())
-            {
-                // get the code of the previous photo
-                DbCommand cmd = conn.CreateCommand();
-                cmd.CommandText = "INSERT INTO Tests_Questions" +
-                    " (IdTest, IdQuestion)" +
-                    " Values" +
-                    " (" + Test.IdTest + "," + Question.IdQuestion + ")" +
-                    "; ";
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-        internal void SaveStudentsAnswer(Student Student, Test Test, Answer Answer,
-            bool StudentsBoolAnswer, string StudentsTextAnswer)
-        {
-            // TODO put this UI matter into form's code 
-            //////////if (Student == null)
-            //////////{
-            //////////    MessageBox.Show("Scegliere un allievo");
-            //////////    return; 
-            //////////}
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                // find if an answer has already been given
-                int? IdStudentsAnswer = StudentHasAnswered(Answer.IdAnswer, Test.IdTest, Student.IdStudent);
-                if (IdStudentsAnswer != null)
-                {   // update answer
-                    cmd.CommandText = "UPDATE StudentsAnswers" +
-                    " SET idStudent=" +  SqlVal.SqlInt(Student.IdStudent) + "," +
-                    "idAnswer=" +  SqlVal.SqlInt(Answer.IdAnswer) + "," +
-                    "studentsBoolAnswer=" + SqlVal.SqlBool(StudentsBoolAnswer) + "," +
-                    "studentsTextAnswer='" + SqlVal.SqlString(StudentsTextAnswer) + "'," +
-                    "IdTest=" +  SqlVal.SqlInt(Test.IdTest) +
-                    "" +
-                    " WHERE IdStudentsAnswer=" + Answer.IdAnswer +
-                    ";";
-                }
-                else
-                {   // create answer
-                    int nextId = dl.NextKey("StudentsAnswers", "IdStudentsAnswer");
-
-                    cmd.CommandText = "INSERT INTO StudentsAnswers " +
-                    "(idStudentsAnswer,idStudent,idAnswer,studentsBoolAnswer," +
-                    "studentsTextAnswer,IdTest" +
-                    ")" +
-                    "Values " +
-                    "(" + nextId + "," +  SqlVal.SqlInt(Student.IdStudent) + "," +
-                     SqlVal.SqlInt(Answer.IdAnswer) + "," + SqlVal.SqlBool(StudentsBoolAnswer) + ",'" +
-                    SqlVal.SqlString(StudentsTextAnswer) + "'," +
-                     SqlVal.SqlInt(Test.IdTest) +
-                    ");";
-                }
-                cmd.ExecuteNonQuery();
-                cmd.Dispose();
-            }
-        }
-        private int? StudentHasAnswered(int? IdAnswer, int? IdTest, int? IdStudent)
-        {
-            int? key;
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT idStudentsAnswer" +
-                    " FROM StudentsAnswers" +
-                    " WHERE idStudent=" + IdStudent +
-                    " AND IdTest=" + IdTest + "" +
-                    " AND IdAnswer=" + IdAnswer + "" +
-                    ";";
-                cmd.CommandText = query;
-                //idStudentsAnswer cmd.ExecuteScalar() != null;
-                key = (int?)cmd.ExecuteScalar();
-            }
-            return key;
-        }
-        internal List<StudentsAnswer> GetAllAnswersOfAStudentToAQuestionOfThisTest(
-            int? IdStudent, int? IdQuestion, int? IdTest)
-        {
-            List<StudentsAnswer> list = new List<StudentsAnswer>(); 
-            using (DbConnection conn = dl.Connect())
-            {
-                DbCommand cmd = conn.CreateCommand();
-                string query = "SELECT *" +
-                    " FROM StudentsAnswers" +
-                    " JOIN Answers ON Answers.idAnswer = StudentsAnswers.idAnswer" +
-                    " JOIN Questions ON Questions.IdQuestion = Answers.IdQuestion" +
-                    " JOIN Tests_Questions ON Questions.IdQuestion = Tests_Questions.IdQuestion" +
-                    " WHERE StudentsAnswers.idStudent=" + IdStudent +
-                    " AND Questions.IdQuestion=" + IdQuestion + "" +
-                    " AND Tests_Questions.IdTest=" + IdTest + "" +
-                    ";";
-
-                cmd.CommandText = query;
-                DbDataReader dRead;
-                dRead = cmd.ExecuteReader();
-
-                while (dRead.Read())
-                {
-                    StudentsAnswer a = GetStudentsAnswerFromRow(dRead);
-                    list.Add(a);
-                }
-            }
-            return list;
-        }
-
-        private StudentsAnswer GetStudentsAnswerFromRow(DbDataReader Row)
-        {
-            StudentsAnswer a = new StudentsAnswer();
-            a.IdAnswer = SafeDb.SafeInt(Row["IdAnswer"]);
-            a.IdStudent = SafeDb.SafeInt(Row["IdStudent"]);
-            a.IdStudentsAnswer = SafeDb.SafeInt(Row["IdStudentsAnswer"]);
-            a.IdTest = SafeDb.SafeInt(Row["IdTest"]);
-            a.StudentsBoolAnswer = SafeDb.SafeBool(Row["StudentsBoolAnswer"]);
-            a.StudentsTextAnswer = SafeDb.SafeString(Row["StudentsTextAnswer"]);
-
-            return a;
-        }
-
-        
-       
         
         #endregion
     }
