+++ MpttTreeManagement/TreeMpttDb.cs
@@ -64,11 +64,11 @@ namespace gamon.TreeMptt
                     {
                         if (t.Id != null && t.Id > 1)
                         {
-                            db.UpdateTopic(t, conn);
+                            dl.UpdateTopic(t, conn);
                         }
                         else
                         {
-                            db.InsertTopic(t, conn);
+                            dl.InsertTopic(t, conn);
                         }
                     }
                     if (!CommonsWinForms.BackgroundCanStillSaveTopicsTree)
@@ -135,7 +135,7 @@ namespace gamon.TreeMptt
                 DbDataReader dRead = cmd.ExecuteReader();
                 while (dRead.Read())
                 {
-                    Topic t = db.GetTopicFromRow(dRead);
+                    Topic t = dl.GetTopicFromRow(dRead);
                     l.Add(t);
                 }
                 dRead.Dispose();
@@ -158,7 +158,7 @@ namespace gamon.TreeMptt
                 DbDataReader dRead = cmd.ExecuteReader();
                 while (dRead.Read())
                 {
-                    Topic t = db.GetTopicFromRow(dRead);
+                    Topic t = dl.GetTopicFromRow(dRead);
                     l.Add(t);
                 }
                 dRead.Dispose();
@@ -254,7 +254,7 @@ namespace gamon.TreeMptt
 
                 while (dRead.Read())
                 {
-                    Topic t = db.GetTopicFromRow(dRead);
+                    Topic t = dl.GetTopicFromRow(dRead);
                     lt.Add(t);
                 }
                 dRead.Dispose();
@@ -287,7 +287,7 @@ namespace gamon.TreeMptt
                 // if CurrentNode is a new node, then we create it in the database, 
                 // so that it will have its Id. It will be saved with correct data 
                 // in the following because new and old values will differ 
-                ct.Id = db.CreateNewTopic(ct);
+                ct.Id = dl.CreateNewTopic(ct);
             }
             int brotherNo = 1;
             foreach (TreeNode sonNode in CurrentNode.Nodes)
@@ -360,7 +360,7 @@ namespace gamon.TreeMptt
             DbDataReader dRead = cmd.ExecuteReader();
             while (dRead.Read())
             {
-                Topic t = db.GetTopicFromRow(dRead);
+                Topic t = dl.GetTopicFromRow(dRead);
                 lt.Add(t);
             }
             dRead.Dispose();
@@ -392,7 +392,7 @@ namespace gamon.TreeMptt
                 DbDataReader dRead = cmd.ExecuteReader();
                 while (dRead.Read())
                 {
-                    Topic t = db.GetTopicFromRow(dRead);
+                    Topic t = dl.GetTopicFromRow(dRead);
                     l.Add(t);
                 }
                 dRead.Dispose();
@@ -443,7 +443,7 @@ namespace gamon.TreeMptt
                 dRead = cmd.ExecuteReader();
                 while (dRead.Read())
                 {
-                    Topic t = db.GetTopicFromRow(dRead);
+                    Topic t = dl.GetTopicFromRow(dRead);
                     found.Add(t);
                 }
                 dRead.Dispose();
