+++ SchoolGrades/frmTestManagement.cs
@@ -15,12 +15,13 @@ namespace SchoolGrades
     {
         DbAndBusiness db;
         List<Test> listTests = new List<Test>();
-        Test currentTest; 
-
+        Test currentTest;
+        DataLayer dl;
         public frmTestManagement()
         {
             InitializeComponent();
             db = new DbAndBusiness(Commons.PathAndFileDatabase);
+            dl = new DataLayer();
         }
 
         private void frmTest_Load(object sender, EventArgs e)
@@ -31,7 +32,7 @@ namespace SchoolGrades
             currentTest = db.GetTest(1);
             // !!!!!!!!!!!
 
-            List<QuestionType> lq = db.GetListQuestionTypes(true);
+            List<QuestionType> lq = dl.GetListQuestionTypes(true);
             cmbQuestionTypes.DisplayMember = "Name";
             cmbQuestionTypes.ValueMember = "idQuestionType";
             cmbQuestionTypes.DataSource = lq;
@@ -62,7 +63,7 @@ namespace SchoolGrades
 
             if (currentTest.IdTest != 0)
             {
-                dgwQuestions.DataSource = db.GetAllQuestionsOfATest(currentTest.IdTest); 
+                dgwQuestions.DataSource = dl.GetAllQuestionsOfATest(currentTest.IdTest); 
             }
             // !!!! put the rest of the data !!!!
         }
@@ -103,7 +104,7 @@ namespace SchoolGrades
             scelta.ShowDialog();
             if (scelta.ChosenQuestion != null && scelta.ChosenQuestion.IdQuestion != 0)
             {
-                db.AddQuestionToTest(currentTest, scelta.ChosenQuestion);
+                dl.AddQuestionToTest(currentTest, scelta.ChosenQuestion);
             }
             scelta.Dispose();
         }
